LayoutTests/imported/w3c/web-platform-tests/css/css-values/calc-size/animation/interpolate-size-logical-properties-interpolation-expected.txt


FAIL CSS Transitions: property <block-size> from [max-content] to [25px] at (-0.3) should be [90px] assert_equals: expected "90px " but got "25px "
FAIL CSS Transitions: property <block-size> from [max-content] to [25px] at (0) should be [75px] assert_equals: expected "75px " but got "25px "
FAIL CSS Transitions: property <block-size> from [max-content] to [25px] at (0.3) should be [60px] assert_equals: expected "60px " but got "25px "
FAIL CSS Transitions: property <block-size> from [max-content] to [25px] at (0.6) should be [45px] assert_equals: expected "45px " but got "25px "
PASS CSS Transitions: property <block-size> from [max-content] to [25px] at (1) should be [25px]
FAIL CSS Transitions: property <block-size> from [max-content] to [25px] at (1.5) should be [0px] assert_equals: expected "0px " but got "25px "
FAIL CSS Transitions with transition: all: property <block-size> from [max-content] to [25px] at (-0.3) should be [90px] assert_equals: expected "90px " but got "25px "
FAIL CSS Transitions with transition: all: property <block-size> from [max-content] to [25px] at (0) should be [75px] assert_equals: expected "75px " but got "25px "
FAIL CSS Transitions with transition: all: property <block-size> from [max-content] to [25px] at (0.3) should be [60px] assert_equals: expected "60px " but got "25px "
FAIL CSS Transitions with transition: all: property <block-size> from [max-content] to [25px] at (0.6) should be [45px] assert_equals: expected "45px " but got "25px "
PASS CSS Transitions with transition: all: property <block-size> from [max-content] to [25px] at (1) should be [25px]
FAIL CSS Transitions with transition: all: property <block-size> from [max-content] to [25px] at (1.5) should be [0px] assert_equals: expected "0px " but got "25px "
FAIL CSS Animations: property <block-size> from [max-content] to [25px] at (-0.3) should be [90px] assert_equals: expected "90px " but got "75px "
PASS CSS Animations: property <block-size> from [max-content] to [25px] at (0) should be [75px]
FAIL CSS Animations: property <block-size> from [max-content] to [25px] at (0.3) should be [60px] assert_equals: expected "60px " but got "75px "
FAIL CSS Animations: property <block-size> from [max-content] to [25px] at (0.6) should be [45px] assert_equals: expected "45px " but got "25px "
PASS CSS Animations: property <block-size> from [max-content] to [25px] at (1) should be [25px]
FAIL CSS Animations: property <block-size> from [max-content] to [25px] at (1.5) should be [0px] assert_equals: expected "0px " but got "25px "
FAIL Web Animations: property <block-size> from [max-content] to [25px] at (-0.3) should be [90px] assert_equals: expected "90px " but got "75px "
PASS Web Animations: property <block-size> from [max-content] to [25px] at (0) should be [75px]
FAIL Web Animations: property <block-size> from [max-content] to [25px] at (0.3) should be [60px] assert_equals: expected "60px " but got "75px "
FAIL Web Animations: property <block-size> from [max-content] to [25px] at (0.6) should be [45px] assert_equals: expected "45px " but got "25px "
PASS Web Animations: property <block-size> from [max-content] to [25px] at (1) should be [25px]
FAIL Web Animations: property <block-size> from [max-content] to [25px] at (1.5) should be [0px] assert_equals: expected "0px " but got "25px "
FAIL CSS Transitions: property <inline-size> from [max-content] to neutral at (-0.3) should be [132.5px] assert_equals: expected "132.5px " but got "100px "
FAIL CSS Transitions: property <inline-size> from [max-content] to neutral at (0) should be [125px] assert_equals: expected "125px " but got "100px "
FAIL CSS Transitions: property <inline-size> from [max-content] to neutral at (0.3) should be [117.5px] assert_equals: expected "117.5px " but got "100px "
FAIL CSS Transitions: property <inline-size> from [max-content] to neutral at (0.6) should be [110px] assert_equals: expected "110px " but got "100px "
PASS CSS Transitions: property <inline-size> from [max-content] to neutral at (1) should be [100px]
FAIL CSS Transitions: property <inline-size> from [max-content] to neutral at (1.5) should be [87.5px] assert_equals: expected "87.5px " but got "100px "
FAIL CSS Transitions with transition: all: property <inline-size> from [max-content] to neutral at (-0.3) should be [132.5px] assert_equals: expected "132.5px " but got "100px "
FAIL CSS Transitions with transition: all: property <inline-size> from [max-content] to neutral at (0) should be [125px] assert_equals: expected "125px " but got "100px "
FAIL CSS Transitions with transition: all: property <inline-size> from [max-content] to neutral at (0.3) should be [117.5px] assert_equals: expected "117.5px " but got "100px "
FAIL CSS Transitions with transition: all: property <inline-size> from [max-content] to neutral at (0.6) should be [110px] assert_equals: expected "110px " but got "100px "
PASS CSS Transitions with transition: all: property <inline-size> from [max-content] to neutral at (1) should be [100px]
FAIL CSS Transitions with transition: all: property <inline-size> from [max-content] to neutral at (1.5) should be [87.5px] assert_equals: expected "87.5px " but got "100px "
FAIL CSS Animations: property <inline-size> from [max-content] to neutral at (-0.3) should be [132.5px] assert_equals: expected "132.5px " but got "125px "
PASS CSS Animations: property <inline-size> from [max-content] to neutral at (0) should be [125px]
FAIL CSS Animations: property <inline-size> from [max-content] to neutral at (0.3) should be [117.5px] assert_equals: expected "117.5px " but got "125px "
FAIL CSS Animations: property <inline-size> from [max-content] to neutral at (0.6) should be [110px] assert_equals: expected "110px " but got "100px "
PASS CSS Animations: property <inline-size> from [max-content] to neutral at (1) should be [100px]
FAIL CSS Animations: property <inline-size> from [max-content] to neutral at (1.5) should be [87.5px] assert_equals: expected "87.5px " but got "100px "
FAIL Web Animations: property <inline-size> from [max-content] to neutral at (-0.3) should be [132.5px] assert_equals: expected "132.5px " but got "125px "
PASS Web Animations: property <inline-size> from [max-content] to neutral at (0) should be [125px]
FAIL Web Animations: property <inline-size> from [max-content] to neutral at (0.3) should be [117.5px] assert_equals: expected "117.5px " but got "125px "
FAIL Web Animations: property <inline-size> from [max-content] to neutral at (0.6) should be [110px] assert_equals: expected "110px " but got "100px "
PASS Web Animations: property <inline-size> from [max-content] to neutral at (1) should be [100px]
FAIL Web Animations: property <inline-size> from [max-content] to neutral at (1.5) should be [87.5px] assert_equals: expected "87.5px " but got "100px "
FAIL CSS Transitions: property <max-block-size> from neutral to [fit-content] at (-0.3) should be [calc-size(fit-content, 130px + size * -0.3)] assert_equals: expected "100px " but got "fit - content "
FAIL CSS Transitions: property <max-block-size> from neutral to [fit-content] at (0) should be [calc-size(fit-content, 100px + size * 0)] assert_equals: expected "100px " but got "fit - content "
FAIL CSS Transitions: property <max-block-size> from neutral to [fit-content] at (0.3) should be [calc-size(fit-content, 70px + size * 0.3)] assert_equals: expected "100px " but got "fit - content "
FAIL CSS Transitions: property <max-block-size> from neutral to [fit-content] at (0.6) should be [calc-size(fit-content, 40px + size * 0.6)] assert_equals: expected "100px " but got "fit - content "
FAIL CSS Transitions: property <max-block-size> from neutral to [fit-content] at (1) should be [calc-size(fit-content, 0px + size * 1)] assert_equals: expected "100px " but got "fit - content "
FAIL CSS Transitions: property <max-block-size> from neutral to [fit-content] at (1.5) should be [calc-size(fit-content, -50px + size * 1.5)] assert_equals: expected "100px " but got "fit - content "
FAIL CSS Transitions with transition: all: property <max-block-size> from neutral to [fit-content] at (-0.3) should be [calc-size(fit-content, 130px + size * -0.3)] assert_equals: expected "100px " but got "fit - content "
FAIL CSS Transitions with transition: all: property <max-block-size> from neutral to [fit-content] at (0) should be [calc-size(fit-content, 100px + size * 0)] assert_equals: expected "100px " but got "fit - content "
FAIL CSS Transitions with transition: all: property <max-block-size> from neutral to [fit-content] at (0.3) should be [calc-size(fit-content, 70px + size * 0.3)] assert_equals: expected "100px " but got "fit - content "
FAIL CSS Transitions with transition: all: property <max-block-size> from neutral to [fit-content] at (0.6) should be [calc-size(fit-content, 40px + size * 0.6)] assert_equals: expected "100px " but got "fit - content "
FAIL CSS Transitions with transition: all: property <max-block-size> from neutral to [fit-content] at (1) should be [calc-size(fit-content, 0px + size * 1)] assert_equals: expected "100px " but got "fit - content "
FAIL CSS Transitions with transition: all: property <max-block-size> from neutral to [fit-content] at (1.5) should be [calc-size(fit-content, -50px + size * 1.5)] assert_equals: expected "100px " but got "fit - content "
PASS CSS Animations: property <max-block-size> from neutral to [fit-content] at (-0.3) should be [calc-size(fit-content, 130px + size * -0.3)]
PASS CSS Animations: property <max-block-size> from neutral to [fit-content] at (0) should be [calc-size(fit-content, 100px + size * 0)]
PASS CSS Animations: property <max-block-size> from neutral to [fit-content] at (0.3) should be [calc-size(fit-content, 70px + size * 0.3)]
FAIL CSS Animations: property <max-block-size> from neutral to [fit-content] at (0.6) should be [calc-size(fit-content, 40px + size * 0.6)] assert_equals: expected "100px " but got "fit - content "
FAIL CSS Animations: property <max-block-size> from neutral to [fit-content] at (1) should be [calc-size(fit-content, 0px + size * 1)] assert_equals: expected "100px " but got "fit - content "
FAIL CSS Animations: property <max-block-size> from neutral to [fit-content] at (1.5) should be [calc-size(fit-content, -50px + size * 1.5)] assert_equals: expected "100px " but got "fit - content "
PASS Web Animations: property <max-block-size> from neutral to [fit-content] at (-0.3) should be [calc-size(fit-content, 130px + size * -0.3)]
PASS Web Animations: property <max-block-size> from neutral to [fit-content] at (0) should be [calc-size(fit-content, 100px + size * 0)]
PASS Web Animations: property <max-block-size> from neutral to [fit-content] at (0.3) should be [calc-size(fit-content, 70px + size * 0.3)]
FAIL Web Animations: property <max-block-size> from neutral to [fit-content] at (0.6) should be [calc-size(fit-content, 40px + size * 0.6)] assert_equals: expected "100px " but got "fit - content "
FAIL Web Animations: property <max-block-size> from neutral to [fit-content] at (1) should be [calc-size(fit-content, 0px + size * 1)] assert_equals: expected "100px " but got "fit - content "
FAIL Web Animations: property <max-block-size> from neutral to [fit-content] at (1.5) should be [calc-size(fit-content, -50px + size * 1.5)] assert_equals: expected "100px " but got "fit - content "
FAIL CSS Transitions: property <max-inline-size> from [max-content] to [100px] at (-0.3) should be [calc-size(max-content, -30px + size * 1.3)] assert_equals: expected "none " but got "100px "
FAIL CSS Transitions: property <max-inline-size> from [max-content] to [100px] at (0) should be [calc-size(max-content, 0px + size * 1)] assert_equals: expected "none " but got "100px "
FAIL CSS Transitions: property <max-inline-size> from [max-content] to [100px] at (0.3) should be [calc-size(max-content, 30px + size * 0.7)] assert_equals: expected "none " but got "100px "
FAIL CSS Transitions: property <max-inline-size> from [max-content] to [100px] at (0.6) should be [calc-size(max-content, 60px + size * 0.4)] assert_equals: expected "none " but got "100px "
FAIL CSS Transitions: property <max-inline-size> from [max-content] to [100px] at (1) should be [calc-size(max-content, 100px + size * 0)] assert_equals: expected "none " but got "100px "
FAIL CSS Transitions: property <max-inline-size> from [max-content] to [100px] at (1.5) should be [calc-size(max-content, 150px + size * -0.5)] assert_equals: expected "none " but got "100px "
FAIL CSS Transitions with transition: all: property <max-inline-size> from [max-content] to [100px] at (-0.3) should be [calc-size(max-content, -30px + size * 1.3)] assert_equals: expected "none " but got "100px "
FAIL CSS Transitions with transition: all: property <max-inline-size> from [max-content] to [100px] at (0) should be [calc-size(max-content, 0px + size * 1)] assert_equals: expected "none " but got "100px "
FAIL CSS Transitions with transition: all: property <max-inline-size> from [max-content] to [100px] at (0.3) should be [calc-size(max-content, 30px + size * 0.7)] assert_equals: expected "none " but got "100px "
FAIL CSS Transitions with transition: all: property <max-inline-size> from [max-content] to [100px] at (0.6) should be [calc-size(max-content, 60px + size * 0.4)] assert_equals: expected "none " but got "100px "
FAIL CSS Transitions with transition: all: property <max-inline-size> from [max-content] to [100px] at (1) should be [calc-size(max-content, 100px + size * 0)] assert_equals: expected "none " but got "100px "
FAIL CSS Transitions with transition: all: property <max-inline-size> from [max-content] to [100px] at (1.5) should be [calc-size(max-content, 150px + size * -0.5)] assert_equals: expected "none " but got "100px "
FAIL CSS Animations: property <max-inline-size> from [max-content] to [100px] at (-0.3) should be [calc-size(max-content, -30px + size * 1.3)] assert_equals: expected "none " but got "max - content "
FAIL CSS Animations: property <max-inline-size> from [max-content] to [100px] at (0) should be [calc-size(max-content, 0px + size * 1)] assert_equals: expected "none " but got "max - content "
FAIL CSS Animations: property <max-inline-size> from [max-content] to [100px] at (0.3) should be [calc-size(max-content, 30px + size * 0.7)] assert_equals: expected "none " but got "max - content "
FAIL CSS Animations: property <max-inline-size> from [max-content] to [100px] at (0.6) should be [calc-size(max-content, 60px + size * 0.4)] assert_equals: expected "none " but got "100px "
FAIL CSS Animations: property <max-inline-size> from [max-content] to [100px] at (1) should be [calc-size(max-content, 100px + size * 0)] assert_equals: expected "none " but got "100px "
FAIL CSS Animations: property <max-inline-size> from [max-content] to [100px] at (1.5) should be [calc-size(max-content, 150px + size * -0.5)] assert_equals: expected "none " but got "100px "
FAIL Web Animations: property <max-inline-size> from [max-content] to [100px] at (-0.3) should be [calc-size(max-content, -30px + size * 1.3)] assert_equals: expected "none " but got "max - content "
FAIL Web Animations: property <max-inline-size> from [max-content] to [100px] at (0) should be [calc-size(max-content, 0px + size * 1)] assert_equals: expected "none " but got "max - content "
FAIL Web Animations: property <max-inline-size> from [max-content] to [100px] at (0.3) should be [calc-size(max-content, 30px + size * 0.7)] assert_equals: expected "none " but got "max - content "
FAIL Web Animations: property <max-inline-size> from [max-content] to [100px] at (0.6) should be [calc-size(max-content, 60px + size * 0.4)] assert_equals: expected "none " but got "100px "
FAIL Web Animations: property <max-inline-size> from [max-content] to [100px] at (1) should be [calc-size(max-content, 100px + size * 0)] assert_equals: expected "none " but got "100px "
FAIL Web Animations: property <max-inline-size> from [max-content] to [100px] at (1.5) should be [calc-size(max-content, 150px + size * -0.5)] assert_equals: expected "none " but got "100px "
FAIL CSS Transitions: property <min-block-size> from [200px] to [min-content] at (-0.3) should be [calc-size(min-content, 260px + size * -0.3)] assert_equals: expected "0px " but got "min - content "
FAIL CSS Transitions: property <min-block-size> from [200px] to [min-content] at (0) should be [calc-size(min-content, 200px + size * 0)] assert_equals: expected "0px " but got "min - content "
FAIL CSS Transitions: property <min-block-size> from [200px] to [min-content] at (0.3) should be [calc-size(min-content, 140px + size * 0.3)] assert_equals: expected "0px " but got "min - content "
FAIL CSS Transitions: property <min-block-size> from [200px] to [min-content] at (0.6) should be [calc-size(min-content, 80px + size * 0.6)] assert_equals: expected "0px " but got "min - content "
FAIL CSS Transitions: property <min-block-size> from [200px] to [min-content] at (1) should be [calc-size(min-content, 0px + size * 1)] assert_equals: expected "0px " but got "min - content "
FAIL CSS Transitions: property <min-block-size> from [200px] to [min-content] at (1.5) should be [calc-size(min-content, -100px + size * 1.5)] assert_equals: expected "0px " but got "min - content "
FAIL CSS Transitions with transition: all: property <min-block-size> from [200px] to [min-content] at (-0.3) should be [calc-size(min-content, 260px + size * -0.3)] assert_equals: expected "0px " but got "min - content "
FAIL CSS Transitions with transition: all: property <min-block-size> from [200px] to [min-content] at (0) should be [calc-size(min-content, 200px + size * 0)] assert_equals: expected "0px " but got "min - content "
FAIL CSS Transitions with transition: all: property <min-block-size> from [200px] to [min-content] at (0.3) should be [calc-size(min-content, 140px + size * 0.3)] assert_equals: expected "0px " but got "min - content "
FAIL CSS Transitions with transition: all: property <min-block-size> from [200px] to [min-content] at (0.6) should be [calc-size(min-content, 80px + size * 0.6)] assert_equals: expected "0px " but got "min - content "
FAIL CSS Transitions with transition: all: property <min-block-size> from [200px] to [min-content] at (1) should be [calc-size(min-content, 0px + size * 1)] assert_equals: expected "0px " but got "min - content "
FAIL CSS Transitions with transition: all: property <min-block-size> from [200px] to [min-content] at (1.5) should be [calc-size(min-content, -100px + size * 1.5)] assert_equals: expected "0px " but got "min - content "
FAIL CSS Animations: property <min-block-size> from [200px] to [min-content] at (-0.3) should be [calc-size(min-content, 260px + size * -0.3)] assert_equals: expected "0px " but got "200px "
FAIL CSS Animations: property <min-block-size> from [200px] to [min-content] at (0) should be [calc-size(min-content, 200px + size * 0)] assert_equals: expected "0px " but got "200px "
FAIL CSS Animations: property <min-block-size> from [200px] to [min-content] at (0.3) should be [calc-size(min-content, 140px + size * 0.3)] assert_equals: expected "0px " but got "200px "
FAIL CSS Animations: property <min-block-size> from [200px] to [min-content] at (0.6) should be [calc-size(min-content, 80px + size * 0.6)] assert_equals: expected "0px " but got "min - content "
FAIL CSS Animations: property <min-block-size> from [200px] to [min-content] at (1) should be [calc-size(min-content, 0px + size * 1)] assert_equals: expected "0px " but got "min - content "
FAIL CSS Animations: property <min-block-size> from [200px] to [min-content] at (1.5) should be [calc-size(min-content, -100px + size * 1.5)] assert_equals: expected "0px " but got "min - content "
FAIL Web Animations: property <min-block-size> from [200px] to [min-content] at (-0.3) should be [calc-size(min-content, 260px + size * -0.3)] assert_equals: expected "0px " but got "200px "
FAIL Web Animations: property <min-block-size> from [200px] to [min-content] at (0) should be [calc-size(min-content, 200px + size * 0)] assert_equals: expected "0px " but got "200px "
FAIL Web Animations: property <min-block-size> from [200px] to [min-content] at (0.3) should be [calc-size(min-content, 140px + size * 0.3)] assert_equals: expected "0px " but got "200px "
FAIL Web Animations: property <min-block-size> from [200px] to [min-content] at (0.6) should be [calc-size(min-content, 80px + size * 0.6)] assert_equals: expected "0px " but got "min - content "
FAIL Web Animations: property <min-block-size> from [200px] to [min-content] at (1) should be [calc-size(min-content, 0px + size * 1)] assert_equals: expected "0px " but got "min - content "
FAIL Web Animations: property <min-block-size> from [200px] to [min-content] at (1.5) should be [calc-size(min-content, -100px + size * 1.5)] assert_equals: expected "0px " but got "min - content "
FAIL CSS Transitions: property <min-inline-size> from neutral to [100px] at (-0.3) should be [calc-size(auto, -30px + size * 1.3)] assert_equals: expected "0px " but got "100px "
FAIL CSS Transitions: property <min-inline-size> from neutral to [100px] at (0) should be [calc-size(auto, 0px + size * 1)] assert_equals: expected "0px " but got "100px "
FAIL CSS Transitions: property <min-inline-size> from neutral to [100px] at (0.3) should be [calc-size(auto, 30px + size * 0.7)] assert_equals: expected "0px " but got "100px "
FAIL CSS Transitions: property <min-inline-size> from neutral to [100px] at (0.6) should be [calc-size(auto, 60px + size * 0.4)] assert_equals: expected "0px " but got "100px "
FAIL CSS Transitions: property <min-inline-size> from neutral to [100px] at (1) should be [calc-size(auto, 100px + size * 0)] assert_equals: expected "0px " but got "100px "
FAIL CSS Transitions: property <min-inline-size> from neutral to [100px] at (1.5) should be [calc-size(auto, 150px + size * -0.5)] assert_equals: expected "0px " but got "100px "
FAIL CSS Transitions with transition: all: property <min-inline-size> from neutral to [100px] at (-0.3) should be [calc-size(auto, -30px + size * 1.3)] assert_equals: expected "0px " but got "100px "
FAIL CSS Transitions with transition: all: property <min-inline-size> from neutral to [100px] at (0) should be [calc-size(auto, 0px + size * 1)] assert_equals: expected "0px " but got "100px "
FAIL CSS Transitions with transition: all: property <min-inline-size> from neutral to [100px] at (0.3) should be [calc-size(auto, 30px + size * 0.7)] assert_equals: expected "0px " but got "100px "
FAIL CSS Transitions with transition: all: property <min-inline-size> from neutral to [100px] at (0.6) should be [calc-size(auto, 60px + size * 0.4)] assert_equals: expected "0px " but got "100px "
FAIL CSS Transitions with transition: all: property <min-inline-size> from neutral to [100px] at (1) should be [calc-size(auto, 100px + size * 0)] assert_equals: expected "0px " but got "100px "
FAIL CSS Transitions with transition: all: property <min-inline-size> from neutral to [100px] at (1.5) should be [calc-size(auto, 150px + size * -0.5)] assert_equals: expected "0px " but got "100px "
PASS CSS Animations: property <min-inline-size> from neutral to [100px] at (-0.3) should be [calc-size(auto, -30px + size * 1.3)]
PASS CSS Animations: property <min-inline-size> from neutral to [100px] at (0) should be [calc-size(auto, 0px + size * 1)]
PASS CSS Animations: property <min-inline-size> from neutral to [100px] at (0.3) should be [calc-size(auto, 30px + size * 0.7)]
FAIL CSS Animations: property <min-inline-size> from neutral to [100px] at (0.6) should be [calc-size(auto, 60px + size * 0.4)] assert_equals: expected "0px " but got "100px "
FAIL CSS Animations: property <min-inline-size> from neutral to [100px] at (1) should be [calc-size(auto, 100px + size * 0)] assert_equals: expected "0px " but got "100px "
FAIL CSS Animations: property <min-inline-size> from neutral to [100px] at (1.5) should be [calc-size(auto, 150px + size * -0.5)] assert_equals: expected "0px " but got "100px "
PASS Web Animations: property <min-inline-size> from neutral to [100px] at (-0.3) should be [calc-size(auto, -30px + size * 1.3)]
PASS Web Animations: property <min-inline-size> from neutral to [100px] at (0) should be [calc-size(auto, 0px + size * 1)]
PASS Web Animations: property <min-inline-size> from neutral to [100px] at (0.3) should be [calc-size(auto, 30px + size * 0.7)]
FAIL Web Animations: property <min-inline-size> from neutral to [100px] at (0.6) should be [calc-size(auto, 60px + size * 0.4)] assert_equals: expected "0px " but got "100px "
FAIL Web Animations: property <min-inline-size> from neutral to [100px] at (1) should be [calc-size(auto, 100px + size * 0)] assert_equals: expected "0px " but got "100px "
FAIL Web Animations: property <min-inline-size> from neutral to [100px] at (1.5) should be [calc-size(auto, 150px + size * -0.5)] assert_equals: expected "0px " but got "100px "
